+++ imports/plugins/core/templates/client/templates/settings.js
